Refactor out products list observer in DomainManagement.Email
Part of #8760

The <EmailData /> component relies on the sites list and I have been
trying to extract that dependency. Getting there runs deep however and
this PR is one in a line of split-up patches to get us there.

This PR removes the observer mixin on the component and replaces it with
the `<QueryProducts />` component.